refactor: calcNormalizedVolume 분리
AnalyserNode에서 정규화된 볼륨을 계산하는 calcNormalizedVolume를
분리했습니다.
